asm: add empty attribute group definition if missing
Note, this is not a pretty fix, but we most likely do
what the LLVM parser is doing to handle these cases.

The fix really belongs upstream, where LLVM IR assembly
files with missing attribute group definitions should
not be created in the first place, and when found, an
error should be reported. Well, that would be in an
idealistic world at least.

Fixes #37.